src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=src/requirements.txt requirements/common.txt
#
asgiref==3.2.10           # via -r requirements/common.txt, django
bootstrap-admin==0.4.4    # via -r requirements/common.txt
certifi==2020.6.20        # via -r requirements/common.txt, requests
chardet==3.0.4            # via -r requirements/common.txt, requests
django-admin-sortable2==0.7.6  # via -r requirements/common.txt
django-appconf==1.0.4     # via -r requirements/common.txt, django-compressor
django-bootstrap3==14.1.0  # via -r requirements/common.txt
django-ckeditor==5.9.0    # via -r requirements/common.txt
django-compressor==2.4    # via -r requirements/common.txt
django-contrib-comments==1.9.2  # via -r requirements/common.txt
django-extensions==3.0.2  # via -r requirements/common.txt
django-js-asset==1.2.2    # via -r requirements/common.txt, django-ckeditor
django-multisite==1.8.0   # via -r requirements/common.txt
django==3.0.8             # via -r requirements/common.txt, django-admin-sortable2, django-appconf, django-bootstrap3, django-contrib-comments
gunicorn==20.0.4          # via -r requirements/common.txt
idna==2.10                # via -r requirements/common.txt, requests, tldextract
psycopg2-binary==2.8.5    # via -r requirements/common.txt
python-dotenv==0.14.0     # via -r requirements/common.txt
pytz==2020.1              # via -r requirements/common.txt, django
rcssmin==1.0.6            # via -r requirements/common.txt, django-compressor
requests-file==1.5.1      # via -r requirements/common.txt, tldextract
requests==2.24.0          # via -r requirements/common.txt, requests-file, tldextract
rjsmin==1.1.0             # via -r requirements/common.txt, django-compressor
six==1.15.0               # via -r requirements/common.txt, django-compressor, django-contrib-comments, django-extensions, requests-file
sqlparse==0.3.1           # via -r requirements/common.txt, django
tldextract==2.2.2         # via -r requirements/common.txt, django-multisite
urllib3==1.25.9           # via -r requirements/common.txt, requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
